+++ src/AxialMovement.h
@@ -24,7 +24,7 @@ class AxialMovement {
   AxialMovement();
   ~AxialMovement();
 
-  void update(double rotation_deg, int *shoulder_target_deg);
+  void update(double rotation_deg, double *shoulder_target_deg);
 
   void loadMove(std::vector<motionmanager::Movement_t> movement_data);
   void startMovement();
